test/cli: move addon tests from test-other.py to new test-addons.py
